# gh-95534: Improve gzip reading speed by 10%