--- a/Misc/NEWS.d/next/Library/2022-09-30-09-22-37.gh-issue-95534.ndEfPj.rst
+++ b/Misc/NEWS.d/next/Library/2022-09-30-09-22-37.gh-issue-95534.ndEfPj.rst
@@ -1,1 +1,1 @@
-`gzip.GzipFile` reads files 10% faster.
+:meth:`gzip.GzipFile` reads files 10% faster.