--- a/Modules/zlibmodule.c
+++ b/Modules/zlibmodule.c
@@ -1524,7 +1524,6 @@ decompress_buf(ZlibDecompressor *self, Py_ssize_t max_length)
             Py_BEGIN_ALLOW_THREADS
             err = inflate(&self->zst, Z_SYNC_FLUSH);
             Py_END_ALLOW_THREADS
-
             switch (err) {
             case Z_OK:            /* fall through */
             case Z_BUF_ERROR:     /* fall through */
@@ -1558,14 +1557,14 @@ decompress_buf(ZlibDecompressor *self, Py_ssize_t max_length)
     self->avail_in_real += self->zst.avail_in;
 
     if (_PyBytes_Resize(&RetVal, self->zst.next_out -
-                        (uint8_t *)PyBytes_AS_STRING(RetVal)) != 0)
+                        (uint8_t *)PyBytes_AS_STRING(RetVal)) != 0) {
         goto error;
+    }
 
     goto success;
 error:
-    RetVal = NULL;
-success:
     Py_CLEAR(RetVal);
+success:
     return RetVal;
 }